activity_main.xml was not updated in earlier commit.
Individual styles are added to Views(Buttons and TextView).
These styles include.
Different Styles are defined in themes/styles.xml for Button  TextColor,TextStyle and TextSize  and  also Text View .
For Main LinearLayout also A style is defines in themes/styles.xml which change gravity and margin of the layout.